+++ src/api/cast/CastConfiguration.ts
@@ -18,6 +18,15 @@ export interface CastConfiguration {
    * @defaultValue `'manual'`
    */
   strategy?: JoinStrategy;
+
+  /**
+   * A flag to enable experimental pipeline with latest features for Chromecast.
+   *
+   * @defaultValue `false`
+   *
+   * @platform ios
+   */
+  useExperimentalPipeline?: boolean;
 }
 
 /**